+++ extensions/Zay-ES-Net/release-notes.md
@@ -1,6 +1,9 @@
 Zay-ES-Net v1.3.2 (unreleased)
 ------------------
-
+* Fixed HostedEntityData.getStringInfo() to send back a null if the ID doesn't exist
+    for the string rather than sending back -1.  A -1 would have been cached but a
+    null won't be.
+    
 
 Zay-ES-Net v1.3.1 (latest)
 ------------------